+++ changelog.md
@@ -1,5 +1,5 @@
 ## Current
+- Added support for Sunshine gamestream server
+## Previous
 - Added more system data from scanning
 - Certificate and pairing tools now available as pip module
-## Previous
-First official release of this Nvidia gamestream plugins for AKL.